% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_scigraph.R
\name{find_scigraph}
\alias{find_scigraph}
\title{#' Find entities using SciGraph API}
\usage{
find_scigraph(text, include_category = NULL, exclude_category = NULL,
  min_length = 4, longest_only = FALSE, include_abbreviation = FALSE,
  include_acronym = FALSE, include_numbers = FALSE)
}
\arguments{
\item{text}{text to annotate}

\item{include_category}{include category}

\item{exclude_category}{exclude category}

\item{min_length}{minimum length of characters to annotate}

\item{longest_only}{if two or more words overlap, take only the longest one}

\item{include_abbreviation}{consider abbreviations in the text}

\item{include_acronym}{consider acronyms in the text}

\item{include_numbers}{consider numbers in the text}
}
\value{
the results from the search
}
\description{
#' Find entities using SciGraph API
}
\examples{
\dontrun{
find_scigraph(text = 'Marfan syndrome the gene DLL', include_category = 'anatomical entity')
}

}
\author{
Francisco Requena
}
