% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_tagger.R
\name{find_tagger}
\alias{find_tagger}
\title{#' Find entities using Tagger API}
\usage{
find_tagger(text_input, bioconcept = c("gene_species", "chemical",
  "disease", "mutation"))
}
\arguments{
\item{text_input}{text}

\item{bioconcept}{bioconcept}
}
\value{
the results from the search
}
\description{
#' Find entities using Tagger API
}
\examples{
\dontrun{
biolink_api('nlp/annotate, content = 'Marfan syndrome', include_category = 'anatomical entity)
}

}
\author{
Francisco Requena
}
