% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_scigraph.R
\name{find_scigraph}
\alias{find_scigraph}
\title{#' Find entities using SciGraph API}
\usage{
find_scigraph(text, include_category = NULL, exclude_category = NULL,
  min_length = 4, longest_only = FALSE, include_abbreviation = FALSE,
  include_acronym = FALSE, include_numbers = FALSE)
}
\arguments{
\item{text}{text}

\item{include_category}{include category}

\item{exclude_category}{exclude category}

\item{min_length}{minimum length}

\item{longest_only}{longest only}

\item{include_abbreviation}{include abbreviation}

\item{include_acronym}{include acronym}

\item{include_numbers}{include numbers}
}
\value{
the results from the search
}
\description{
#' Find entities using SciGraph API
}
\examples{
\dontrun{
biolink_api('nlp/annotate, content = 'Marfan syndrome', include_category = 'anatomical entity)
}

}
\author{
Francisco Requena
}
